Updated home screen to show heatmap
