+++ client/src/components/ReportObstruction/Upload.jsx
@@ -3,6 +3,8 @@ import Dropzone from "react-dropzone";
 import { LOCATION } from "../../constants/report";
 import { doUploadImage } from "../../redux/report/effects";
 import { Button, Container, Header } from 'semantic-ui-react'
+import { Link } from 'react-router-dom'
+
 
 const Upload = ({ setImgUrl, setStatus }) => {
   const onDrop = (files) => {
@@ -18,7 +20,10 @@ const Upload = ({ setImgUrl, setStatus }) => {
 
   return (
   <Container className="main-container">
+
     <Header as='h1' textAlign="center">Report Obstruction</Header>
+    <Button fluid as={Link} to="/" size='massive'>Home</Button>
+    <br/>
     <Dropzone onDrop={onDrop}>
       {({getRootProps, getInputProps, isDragActive}) => {
         return (
